17 Data Validation

+++ api/Controllers/CommentController.cs
@@ -17,12 +17,18 @@ namespace api.Controllers {
         }
         [HttpGet]
         public async Task<IActionResult> GetAll() {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var comments = await _commentRepository.GetAllAsync();
             var commentDto = comments.Select(func => func.ToCommentDto());
             return Ok(commentDto);
         }
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById([FromRoute] int id) {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var comment = await _commentRepository.GetByIdAsync(id);
             if(comment == null) {
                 return NotFound();
@@ -33,6 +39,9 @@ namespace api.Controllers {
         [HttpPost]
         [Route("{stockId}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto createCommentDto) {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             if(!await _stockRepository.StockExists(stockId)) {
                 return BadRequest("Stock does not exist");
             }
@@ -43,6 +52,9 @@ namespace api.Controllers {
         [HttpPut]
         [Route("{id}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateCommentRequestDto) {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var comment = await _commentRepository.UpdateAsync(id, updateCommentRequestDto.ToCommentFromUpdate());
             if(comment == null) {
                 return NotFound("Comment not found");
@@ -52,6 +64,9 @@ namespace api.Controllers {
         [HttpDelete]
         [Route("{id}")]
         public async Task<IActionResult> Delete([FromRoute] int id) {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var commentModel = await _commentRepository.DeleteAsync(id);
             if(commentModel == null) {
                 return NotFound("comment does not exist");
